begin restructuring project to have common frontend/backend code
