+++ common/tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+        "experimentalDecorators": true,
+        "declaration": true,
+        "outDir": "dist",
+        "sourceMap": true
+    },
+    "include": [
+        "**/*.ts"
+    ],
+    "exclude": [
+        "node_modules",
+        "./dist/**/*"
+    ],
+    "target": "es2017",
+    "module": "commonjs",
+    "lib": [
+        "es2017"
+    ]
+}
